+++ resgen/constants.py
